requirements.txt
huggingface_hub==1.1.2
langchain_core==1.0.5
langchain_openai==1.0.3
python-dotenv==1.2.1
